Split manifest to own file, extract shared money schema (DRY)
Maintainability improvements:

- Move Manifest() and all 6 action schema declarations from square.go
  to manifest.go, reducing square.go from 428 to 171 lines. square.go
  now focuses on the connector struct and HTTP plumbing only.
- Extract duplicated money object schema (amount + currency) into a
  shared moneySchema variable reused by create_order and create_payment.
- Each action manifest is now its own function for easier navigation
  and independent editing in Phase 2.
